refactor Stellar helpers into domain modules
