+++ frontend/lib/stellar/assets.ts
@@ -1,0 +1,30 @@
+import { Asset } from "@stellar/stellar-sdk";
+import { NETWORK } from "../stellarConfig";
+
+export interface KnownAsset {
+  code: string;
+  issuer: string;
+}
+
+export const USDC_ISSUER =
+  process.env.NEXT_PUBLIC_USDC_ISSUER ||
+  "GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN";
+
+export const USDC = new Asset("USDC", USDC_ISSUER);
+
+export const KNOWN_ASSETS: Record<"testnet" | "mainnet", KnownAsset[]> = {
+  testnet: [
+    { code: "USDC", issuer: "GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN" },
+    { code: "AQUA", issuer: "GBBD47IF6LWK7P7MDEVSCWR7DPUWV3NY3DTQEVFL4NAT4AQH3ZLLFLA7" },
+    { code: "yXLM", issuer: "GARDNV3Q7YGT4AKSDF25LT32YSCCW4EV22Y2TV3I2PU2MMXJTEDL5T55" },
+  ],
+  mainnet: [
+    { code: "USDC", issuer: "GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN" },
+    { code: "AQUA", issuer: "GBBD47IF6LWK7P7MDEVSCWR7DPUWV3NY3DTQEVFL4NAT4AQH3ZLLFLA7" },
+    { code: "yXLM", issuer: "GARDNV3Q7YGT4AKSDF25LT32YSCCW4EV22Y2TV3I2PU2MMXJTEDL5T55" },
+  ],
+};
+
+export function getKnownAssets(): KnownAsset[] {
+  return KNOWN_ASSETS[NETWORK];
+}
